+ Option to clip machine names in the machine view for skin creators. See generator_name_clip_coords and effect_name_clip_coords in how-to-skin-psycle.txt. + Sampulse Editor: Implemented the sampulse note mapping editor. Everything is point and click. Multiple selection is possible. (like setting a whole octave by selecting the 12 notemaps, and then the 12 notes with mouse and shift) + Sampulse Editor: Selecting an instrument shows which samples it has associated. Selecting a sample shows which instruments use it. + Sampulse Editor: Deleting an instrument asks if it should delete its associated samples. + Recording or playing the song with looping disabled will now stop when break to line, or jump to sequence go to an earlier position. + Full Implementation of extraticks and ticksperbeat: Changes in song properties dialog, song toolbar and part of code. * Do not stretch master VU graphics in Windows XP. It will not resize correctly on different dpi sizes, but will paint correctly at 96dpi which is what most people use. * Improved the xm song exporter. Fixed sample pitch and exported envelopes. Also exported sampulse instruments now. ! Sampler: Fixes in pitch slide command (it became broken when adding samplerate support) ! Wave editor: Fixes from 1.11 betas ( Crop command broken, copy&paste to new slot, fadein/out of parts of the sample) ! Gear rack: exchanging instruments to an unused position crashed the application.

+++ psycle/src/psycle/host/Sampler.cpp
@@ -872,14 +872,14 @@ namespace psycle
 			{
 				// 0x01 : Pitch Up
 				case 0x01:
-					shift=_voices[voice].effVal*4294967 * _voices[voice]._wave._samplerate/Global::player().SampleRate();
+					shift=static_cast<int64_t>(_voices[voice].effVal)*4294967ll * static_cast<float>(_voices[voice]._wave._samplerate)/Global::player().SampleRate();
 					_voices[voice]._wave._speed+=shift;
 					_resampler.UpdateSpeed(_voices[voice].resampler_data,_voices[voice]._wave._speed);
 				break;
 
 				// 0x02 : Pitch Down
 				case 0x02:
-					shift=_voices[voice].effVal*4294967 * _voices[voice]._wave._samplerate/Global::player().SampleRate();
+					shift=static_cast<int64_t>(_voices[voice].effVal)*4294967ll * static_cast<float>(_voices[voice]._wave._samplerate)/Global::player().SampleRate();
 					_voices[voice]._wave._speed-=shift;
 					if ( _voices[voice]._wave._speed < 0 ) _voices[voice]._wave._speed=0;
 					_resampler.UpdateSpeed(_voices[voice].resampler_data,_voices[voice]._wave._speed);
